Validation working but issues
The form validation is working but only with all single checks and some checks that have multiple properties to validate against.

For ex: If you validate against integer and blank values, the expexted behavior is to tell you that if it's blank, it alerts you it's blank and req. and if you fill in a non integer it will alert you again that the value is not an integer.

However this is not the case. If you enter a non-integer it will not alert you at all because the non integer is thrown, but the non blank check is run next and since it's not blank, it erases the prev error.